Bug Fixes
Bug Fix: The routes dictionary was doing regex matching on the beginning of the string. This caused problems when the same stack handled two paths that started the same - f.e '/' and '/h'.
Bug Fix: The request handler did not handle the case where a query string was specified. Added a test (test_advanced.py) to cover this use-case