Enhancement: Access to the StackInABoxServices
Adds a parameter to the decorator that will cause the decorator to
pass access to the StackInABoxServices declared in the decorator's
parameter list to the test function as a keyword argument by the
name of the value of the parameter.

That is, if the 'access_services' parameter's value is 'stack' it
will add a keyword parameter by the name 'stack' to the  keyword
parameters of the test function that contains a dictionary of the
StackInABoxServices. The name of the service is the keyword to the
dictionary.